+++ tdrs-backend/tdpservice/email/email.py
@@ -1,7 +1,5 @@
 """Wrapper to send emails with Django."""
 
-from tdpservice.email.email_enums import EmailType
-
 from celery import shared_task
 from django.core.exceptions import ValidationError
 from django.core.validators import validate_email
@@ -12,7 +10,7 @@ from django.contrib.admin.models import LogEntry, ContentType, CHANGE
 
 import logging
 
-logger = logging.getLogger(__name__)
+logger = logging.getLogger()
 
 
 def log(msg, logger_context=None, level='info'):
@@ -44,67 +42,6 @@ def log(msg, logger_context=None, level='info'):
             object_id=logger_context['user_id'],
             object_repr=logger_context['user_email']
         )
-
-
-def send_approval_status_update_email(
-    new_approval_status,
-    user,
-    context,
-):
-    """Send an email to a user when their account approval status is updated."""
-    from tdpservice.users.models import AccountApprovalStatusChoices
-
-    recipient_email = user.email
-    logger_context = {
-        'user_id': user.id,
-        'user_email': user.email
-    }
-
-    template_path = None
-    subject = None
-    text_message = None
-
-    log(f"Preparing email to {recipient_email} with status {new_approval_status}", logger_context=logger_context)
-
-    match new_approval_status:
-        case AccountApprovalStatusChoices.INITIAL:
-            # Stubbed for future use
-            return
-
-        case AccountApprovalStatusChoices.ACCESS_REQUEST:
-            template_path = EmailType.ACCESS_REQUEST_SUBMITTED.value
-            subject = 'Access Request Submitted'
-            text_message = 'Your account has been requested.'
-
-        case AccountApprovalStatusChoices.PENDING:
-            # Stubbed for future use
-            return
-
-        case AccountApprovalStatusChoices.APPROVED:
-            template_path = EmailType.REQUEST_APPROVED.value
-            subject = 'Access Request Approved'
-            text_message = 'Your account request has been approved.'
-
-        case AccountApprovalStatusChoices.DENIED:
-            template_path = EmailType.REQUEST_DENIED.value
-            subject = 'Access Request Denied'
-            text_message = 'Your account request has been denied.'
-
-        case AccountApprovalStatusChoices.DEACTIVATED:
-            template_path = EmailType.ACCOUNT_DEACTIVATED.value
-            subject = 'Account is Deactivated'
-            text_message = 'Your account has been deactivated.'
-
-    context.update({'subject': subject})
-
-    automated_email.delay(
-        email_path=template_path,
-        recipient_email=recipient_email,
-        subject=subject,
-        email_context=context,
-        text_message=text_message,
-        logger_context=logger_context
-    )
 
 
 @shared_task
